Add AES256_CTS_HMAC_SHA1_96
This adds the needed cryptographic routines for RFC3962 (etype 18)
AES256_CTS_HMAC_SHA1_96. This was chosen to be added over
RFC8009 AES256_CTS_HMAC_SHA384_192 as while MIT KRB5 claims to
support AES256_CTS_HMAC_SHA384_192, when requesting tickets using
this etype it will never respond to the request. The major
concern with this type is the use of HMAC SHA1, however HMAC SHA1
is not vulnurable to the same pre-image attacks that makes SHA1
collisions trivial. Regardless, we should consider this a stepping
stone to AES256_CTS_HMAC_SHA384_192 support.